WIP ES6: refactor CRS/Projection modules, CRS obj -> CRS.Base
