+++ src/geo/crs/CRS.EPSG3395.js
@@ -1,16 +1,20 @@
+import {Earth} from './CRS.Earth';
+import {Mercator} from '../projection/Projection.Mercator';
+import {Transformation} from '../../geometry/Transformation';
+import {extend} from '../../core/Util';
+
 /*
  * @namespace CRS
  * @crs L.CRS.EPSG3395
  *
  * Rarely used by some commercial tile providers. Uses Elliptical Mercator projection.
  */
-
-L.CRS.EPSG3395 = L.extend({}, L.CRS.Earth, {
+export var EPSG3395 = extend({}, Earth, {
 	code: 'EPSG:3395',
-	projection: L.Projection.Mercator,
+	projection: Mercator,
 
 	transformation: (function () {
-		var scale = 0.5 / (Math.PI * L.Projection.Mercator.R);
-		return new L.Transformation(scale, 0.5, -scale, 0.5);
+		var scale = 0.5 / (Math.PI * Mercator.R);
+		return new Transformation(scale, 0.5, -scale, 0.5);
 	}())
 });
