+++ src/geo/projection/Projection.LonLat.js
@@ -1,3 +1,7 @@
+import {LatLng} from '../LatLng';
+import {Bounds} from '../../geometry/Bounds';
+import {Point} from '../../geometry/Point';
+
 /*
  * @namespace Projection
  * @section
@@ -11,16 +15,14 @@
  * `EPSG:3395` and `Simple` CRS.
  */
 
-L.Projection = {};
-
-L.Projection.LonLat = {
+export var LonLat = {
 	project: function (latlng) {
-		return new L.Point(latlng.lng, latlng.lat);
+		return new Point(latlng.lng, latlng.lat);
 	},
 
 	unproject: function (point) {
-		return new L.LatLng(point.y, point.x);
+		return new LatLng(point.y, point.x);
 	},
 
-	bounds: L.bounds([-180, -90], [180, 90])
+	bounds: new Bounds([-180, -90], [180, 90])
 };
