--- src/geo/projection/Projection.leafdoc
@@ -1,16 +1,0 @@
-
-@class Projection
-
-An object with methods for projecting geographical coordinates of the world onto
-a flat surface (and back). See [Map projection](http://en.wikipedia.org/wiki/Map_projection).
-
-@property bounds: Bounds
-The bounds (specified in CRS units) where the projection is valid
-
-@method project(latlng: LatLng): Point
-Projects geographical coordinates into a 2D point. Only accepts actual `L.LatLng` instances, not arrays.
-
-@method unproject(point: Point): LatLng
-The inverse of `project`. Projects a 2D point into a geographical location. Only accepts actual `L.Point` instances, not arrays.
-
-
